+++ app/features/juz/[juzId]/page.tsx
@@ -9,12 +9,12 @@ import { TranslationPanel } from '@/app/features/surah/[surahId]/_components/Tra
 import { WordLanguagePanel } from '@/app/features/surah/[surahId]/_components/WordLanguagePanel';
 import { Verse as VerseType, TranslationResource, Juz } from '@/types';
 import { getTranslations, getWordTranslations, getVersesByJuz, getJuz } from '@/lib/api';
-import { LANGUAGE_CODES } from '@/lib/languageCodes';
-import type { LanguageCode } from '@/lib/languageCodes';
-import { WORD_LANGUAGE_LABELS } from '@/lib/wordLanguages';
+import { LANGUAGE_CODES } from '@/lib/text/languageCodes';
+import type { LanguageCode } from '@/lib/text/languageCodes';
+import { WORD_LANGUAGE_LABELS } from '@/lib/text/wordLanguages';
 import { useSettings } from '@/app/context/SettingsContext';
 import { useAudio } from '@/app/features/player/context/AudioContext';
-import { buildAudioUrl } from '@/app/features/player/lib/reciters';
+import { buildAudioUrl } from '@/lib/audio/reciters';
 import useSWR from 'swr';
 import useSWRInfinite from 'swr/infinite';
