+++ app/features/tafsir/[surahId]/[ayahId]/_components/TafsirVerse.tsx
@@ -8,14 +8,14 @@ import {
   FaChevronDown,
 } from '@/app/components/shared/SvgIcons';
 import { Verse as VerseType, Translation, Word } from '@/types';
-import type { LanguageCode } from '@/lib/languageCodes';
+import type { LanguageCode } from '@/lib/text/languageCodes';
 import { useAudio } from '@/app/features/player/context/AudioContext';
 import { useSettings } from '@/app/context/SettingsContext';
 import { useState } from 'react';
 import Spinner from '@/app/components/shared/Spinner';
-import { applyTajweed } from '@/lib/tajweed';
-import { getTafsirCached } from '@/app/features/tafsir/lib/tafsirCache';
-import { applyArabicFont } from '@/app/features/tafsir/lib/applyArabicFont';
+import { applyTajweed } from '@/lib/text/tajweed';
+import { getTafsirCached } from '@/lib/tafsir/tafsirCache';
+import { applyArabicFont } from '@/lib/tafsir/applyArabicFont';
 
 interface TafsirVerseProps {
   verse: VerseType;
