+++ lib/tafsir/applyArabicFont.ts
