+++ lib/text/languageCodes.ts
