+++ lib/text/stripHtml.ts
