+++ lib/text/tajweed.ts
