+++ lib/text/wordLanguages.ts
