+++ types/word.ts
@@ -1,4 +1,4 @@
-import type { LanguageCode } from '@/lib/languageCodes';
+import type { LanguageCode } from '@/lib/text/languageCodes';
 
 export interface Word extends Partial<Record<Exclude<LanguageCode, 'id'>, string>> {
   id: number;
